+++ src/main/java/com/mauriciotogneri/botcoin/strategy/buy/BuyStrategy.java
@@ -2,5 +2,5 @@ package com.mauriciotogneri.botcoin.strategy.buy;
 
 public interface BuyStrategy
 {
-    float buy(float price);
+    double buy(double price);
 }
