+++ src/main/java/com/mauriciotogneri/botcoin/strategy/sell/SellStrategy.java
@@ -2,5 +2,5 @@ package com.mauriciotogneri.botcoin.strategy.sell;
 
 public interface SellStrategy
 {
-    float sell(float price);
+    double sell(double price);
 }
